pyproject.toml
[project]
name = "buy-me-coffee-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.0",
    "vyper>=0.4.1",
    "web3>=7.11.0",
]
